refactor: move ElmHeading and ElmFragmentIdentifier components to typography directory
